Removed IncompletePhiNodes list. Modified mapFloatCastToAFValue to handle phi nodes being casted.

+++ llvm/test/Transforms/ErrorAnalysis/LargeBenchmarks/Lorenz/lorenz_simple.c
@@ -9,8 +9,8 @@ double RHO = 12;
 double SIGMA = 10;
 double BETA = 8/3;
 
-void solve_lorenz(double x_0, double y_0, double z_0) {
-  double x_1 = 0, y_1 = 0, z_1 = 0;
+void solve_lorenz(float x_0, float y_0, float z_0) {
+  float x_1 = 0, y_1 = 0, z_1 = 0;
   int n = 10;
 
   char command_filename[] = "lorenz_ode_commands.txt";
@@ -44,18 +44,18 @@ void solve_lorenz(double x_0, double y_0, double z_0) {
     z_0 = z_1 + (x_1*y_1 - BETA*z_1)*0.005;
   }
 
-  printf("x_0 = %0.15f\n", x_0);
-  printf("y_0 = %0.15f\n", y_0);
-  printf("z_0 = %0.15lf\n\n", z_0);
+  printf("x_0 = %0.7f\n", x_0);
+  printf("y_0 = %0.7f\n", y_0);
+  printf("z_0 = %0.7lf\n\n", z_0);
 
   fprintf ( data_unit, "  %d  %14.6g  %14.6g  %14.6g\n",
           2*n, x_0, y_0, z_0 );
 }
 
 int main() {
-  double x_0 = 1.2;
-  double y_0 = 1.3;
-  double z_0 = 1.4;
+  float x_0 = 1.2;
+  float y_0 = 1.3;
+  float z_0 = 1.4;
 //  printf("Enter value of z: ");
 //  scanf("%lf", &z_0);
